requirements.txt
apache-airflow==2.7.0
pytest==7.4.0
black